+++ oneshot_dwdx.cpp
@@ -42,7 +42,7 @@ void oneshot_dwdx(
 	const int n_resi = 3*n_elem;
 	const int n_face = n_elem+1;
 
-	if(n_elem!=x.size()-1) abort();
+	if(n_elem!=(int)x.size()-1) abort();
 	class Flow_data<double> flow_data(n_elem);
     // Inlet flow properties
     double inlet_T = isenT(gam, flo_opts.inlet_total_T, flo_opts.inlet_mach);
